Add InterceptJSInvokeNodes in WasmGC backend
Without them, functional interfaces passed to JavaScript could have
their method not generated, making them not callable from JavaScript
